--- app/src/main/java/com/studentlifemanager/utils/MonthYearPickerDialog.kt
@@ -1,86 +1,0 @@
-package com.studentlifemanager.utils
-
-import android.app.AlertDialog
-import android.app.DatePickerDialog
-import android.app.Dialog
-import android.os.Bundle
-import androidx.fragment.app.DialogFragment
-import com.studentlifemanager.common.R
-import com.studentlifemanager.databinding.DialogMonthYearPickerBinding
-import java.util.Calendar
-import java.util.Date
-
-/**
- * This is the utility class to handle month and year click
- * Here dialog fragment is implemented in order to select month and year
- *
- * @author SandeepK
- * @version 2.0
- * */
-class MonthYearPickerDialog(val date: Date = Date()) : DialogFragment() {
-
-    companion object {
-        private const val MAX_YEAR = 2099
-    }
-
-    private lateinit var binding: DialogMonthYearPickerBinding
-
-    private var listener: DatePickerDialog.OnDateSetListener? = null
-
-    fun setListener(listener: DatePickerDialog.OnDateSetListener?) {
-        this.listener = listener
-    }
-
-    override fun onCreateDialog(savedInstanceState: Bundle?): Dialog {
-        binding = DialogMonthYearPickerBinding.inflate(requireActivity().layoutInflater)
-        val cal: Calendar = Calendar.getInstance().apply { time = date }
-
-        binding.pickerMonth.run {
-            minValue = 0
-            maxValue = 11
-            value = cal.get(Calendar.MONTH)
-            displayedValues = getMonth()
-        }
-
-        binding.pickerYear.run {
-            val year = cal.get(Calendar.YEAR)
-            minValue = year
-            maxValue = MAX_YEAR
-            value = year
-        }
-
-        return AlertDialog.Builder(requireContext())
-            .setTitle(getString(R.string.please_select_the_month))
-            .setView(binding.root)
-            .setPositiveButton(getString(R.string.lbl_ok)) { _, _ ->
-                listener?.onDateSet(
-                    null,
-                    binding.pickerYear.value,
-                    binding.pickerMonth.value,
-                    1
-                )
-            }
-            .setNegativeButton(getString(R.string.lbl_cancel)) { _, _ -> dialog?.cancel() }
-            .create()
-    }
-
-    /**
-     * get months list
-     */
-    fun getMonth(): Array<String> {
-        return arrayOf(
-            getString(R.string.month_january),
-            getString(R.string.month_february),
-            getString(R.string.month_march),
-            getString(R.string.month_april),
-            getString(R.string.month_may),
-            getString(R.string.month_june),
-            getString(R.string.month_july),
-            getString(R.string.month_august),
-            getString(R.string.month_september),
-            getString(R.string.month_october),
-            getString(R.string.month_november),
-            getString(R.string.month_december)
-        )
-    }
-}
