+++ common/src/main/java/com/studentlifemanager/common/utils/Constant.kt
@@ -1,11 +1,9 @@
 package com.studentlifemanager.common.utils
 
 object Constant {
+
     // database
     const val DATABASE_NAME = "app_database"
-
-    var INCOME = 1
-    var EXPENSE = 2
 
     // date format constant
     const val FULL_MONTH = "MMMM"
@@ -13,6 +11,14 @@ object Constant {
     const val DD_MM_YYYY = "dd-MM-yyyy"
     const val DD_MMM_YYYY = "dd MMM yyyy"
     const val YYYY_MM_DD = "yyyy-MM-dd"
+
+    // expense module
+    var INCOME = 1
+    var EXPENSE = 2
+
+    // pin module
+    const val PIN_WEB = 1
+    const val PIN_VIDEO = 2
 
     // app level tags
     const val DATE_PICKER_TAG = "MonthYearPickerDialog"
