+++ expense/src/main/java/com/studentlifemanager/expense/utils/BindableAdapter.kt
@@ -1,4 +1,4 @@
-package com.studentlifemanager.utils
+package com.studentlifemanager.expense.utils
 
 interface BindableAdapter<T> {
     fun setData(data: T)
