--- expense/src/main/java/com/studentlifemanager/expense/utils/Constant.kt
@@ -1,24 +1,0 @@
-package com.studentlifemanager.utils
-
-object Constant {
-
-    var INCOME = 1
-    var EXPENSE = 2
-
-
-    // date format constant
-    const val FULL_MONTH = "MMMM"
-    const val FULL_YEAR = "YYYY"
-    const val DD_MM_YYYY = "dd-MM-yyyy"
-    const val DD_MMM_YYYY = "dd MMM yyyy"
-
-    // app level tags
-    const val DATE_PICKER_TAG = "MonthYearPickerDialog"
-
-    // string
-    const val EMPTY = ""
-
-    // database
-    const val DATABASE_NAME = "app_database"
-
-}
