--- expense/src/main/res/drawable/ic_filter_subject.xml
@@ -1,12 +1,0 @@
-<vector android:height="512dp" android:viewportHeight="20"
-    android:viewportWidth="20" android:width="512dp" xmlns:android="http://schemas.android.com/apk/res/android">
-    <path android:fillColor="#ffffff"
-        android:pathData="M14.32,6.95L5.68,6.95a0.38,0.38 0,0 0,-0.38 0.38,0.37 0.37,0 0,0 0.38,0.37h8.64a0.37,0.37 0,0 0,0.38 -0.37A0.38,0.38 0,0 0,14.32 6.95Z"
-        android:strokeColor="#ffffff" android:strokeWidth="1"/>
-    <path android:fillColor="#ffffff"
-        android:pathData="M12.32,9.64L7.68,9.64a0.38,0.38 0,1 0,0 0.75h4.64a0.38,0.38 0,1 0,0 -0.75Z"
-        android:strokeColor="#ffffff" android:strokeWidth="1"/>
-    <path android:fillColor="#ffffff"
-        android:pathData="M11.32,12.3L8.68,12.3a0.38,0.38 0,1 0,0 0.75h2.64a0.38,0.38 0,1 0,0 -0.75Z"
-        android:strokeColor="#ffffff" android:strokeWidth="1"/>
-</vector>
