--- expense/src/main/res/drawable/ic_plus.xml
@@ -1,10 +1,0 @@
-<vector xmlns:android="http://schemas.android.com/apk/res/android"
-    android:width="512dp"
-    android:height="512dp"
-    android:viewportWidth="512"
-    android:viewportHeight="512">
-  <path
-      android:pathData="M224,112l0,112 -112,-0 -112,-0 0,32 0,32 112,-0 112,-0 0,112 0,112 32,-0 32,-0 0,-112 0,-112 112,-0 112,-0 0,-32 0,-32 -112,-0 -112,-0 0,-112 0,-112 -32,-0 -32,-0 0,112z"
-      android:fillColor="#FFFFFF"
-      android:strokeColor="#00000000"/>
-</vector>
