--- pin/src/main/java/com/studentlifemanager/pin/utility/Constant.kt
@@ -1,7 +1,0 @@
-package com.studentlifemanager.pin.utility
-
-object Constant {
-    const val PIN_WEB = 1
-    const val PIN_VIDEO = 2
-
-}
